deploy: don't change ref and validate checksums
Commit beb026f7 introduced an interesting behaviour. Doing
`rpm-ostree deploy $X` would find the commit associated with the version
tag, fetch it, and set the refspec to that new commit.

This means that whether we deploy an older or newer version, the refspec
would unconditionally get set to that new deployment's commit. We should
instead try to have the refspec always point to the latest known remote
commit, much like git remote refs work.

Another funky behaviour of `rpm-ostree deploy` was that specifying a
csum directly allowed you to jump to *any* commit, regardless of whether
that commit exists on the current branch or not. We tighten that up here
so that we check that the checksum does exist on the current branch.

That behaviour can be useful of course, but we might want to change how
users access it so that we don't get inconsistencies such as `rpm-ostree
status` saying that we're sitting on a specific branch with a specific
commit which doesn't actually belong to that branch.