Run bwrap with fixed environment (PATH + LANG)
This is basically a re-implementation of
https://github.com/GNOME/libglnx/commit/85c9dd5c073a8c0d74c4baa2e4a94f5535984e62
for the same reasons: it makes things work for rpm-ostree running
inside NixOS.  But there are other cases where if e.g. an unprivileged
user runs rpm-ostree and may not have `/usr/sbin` in PATH, we still
want the container to pick it up, etc.

While changing things I nuked the hacky `RPMOSTREE_DEBUG_SCRIPT` env
var, we can re-add it later in a better way if needed.