vmcheck: Ability to run without make and other build tools on host
Installing build tools directly on a host system should be considered
an explicit antipattern for Project Atomic (whether servers or
workstations).  I started on changes to be able to run
`./tests/vmcheck/test.sh` directly, assuming I have vagrant on the
host.

However, I ran into the fact that we need to build the test RPMs
etc inside the VM, not on the host.

Filing this as a start of reworking things, will try to pick it up
later.