WIP: local package install
Ok, I thought this wouldn't be too bad, but the sticky part is a lot of core
operates in terms of `DnfPackage`, which doesn't know how to pull from ostree.
So we end up duplicating/casing a lot of the code paths there.  (Hm, maybe
we could subclass `DnfPackage`?)

It's tempting to just copy the package into `/var/lib/rpm-ostree/local` and
automatically inject a `file://` repo pointing to it, but...