daemon: Do tmprootfs checkout just-in-time if possible
It annoyed me that e.g. `rpm-ostree install nosuchpackage` would spend time
doing a tmprootfs checkout, only to discard it. If the booted deployment's base
is the same as what we would check out, reuse it as long as possible. Only if
everything else succeeded (dependency resolution, downloads, imports), do we go
to check out a copy.

However in the rebase (or upgrade + layering simultaneously) case, we do need to
do a checkout immediately, as the rebase may have changed the OS major version
in `/usr/lib/os-release` which would affect our layering.