core: Also apply mode overrides to symlinks
I originally thought there was a bug here, but missed we were skipping
symlinks earlier above.  See [previous discussion][1].

Now, I'm not aware right now of something this patch actively fixes, but I think
it makes sense to be consistent in our handling of things here with respect to
symlinks.

1: https://github.com/projectatomic/rpm-ostree/pull/561/files/29dd1bd8012f0d34385fb5b1225bf47ccd5b4249..8158dcfb47a112f56c9e6565085d180f359f1b3c#r95017893